requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.2
billiard==3.3.0.23
celery==3.1.24
Django==1.8.2
django-celery==3.2.1
django-ckeditor==5.2.1
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-model-utils==2.6.1
django-redis==4.7.0
djangorestframework==3.5.2
docx==0.2.4
fakeredis==0.8.2
jsonfield==2.0.0
kombu==3.0.37
lxml==3.7.3
MySQL-python==1.2.5
packaging==16.8
Pillow==3.4.2
pyparsing==2.2.0
python-docx==0.8.6
pytz==2016.10
redis==2.10.5
requests==2.13.0
six==1.10.0
